Debug transcribe basic errors
Transcribe has now been tested on a mock server that counts the number
of requests that have come in to all paths and gives that number to the
requestor. This gives Transcribe a session ID and request numbers as
transcripts to see if it's monotonic in the UI textbox. So far so good.

Mic start/stop and start/stop transcription seem to work, as does
submit which sends a request via Fetch and reloads the page on success.