New handshake on each mic start/stop, refactor request index management
Handshake now fires every time the mic is unmuted. [May want to move that
to mute handler to reduce downtime, but for now this is more intuitive.]
Needed to refactor the request index (i.e. session_idx) for correctness
(unseen bug allowed race condition previously with many requests
simultaneously resolved) and to ensure that the current session ID is
accurate.