Fix ordering with mic_idx
`mic_idx` should be compared against `rx_idx` to see if the current request
is about to set the latest response into the text field. Because of setState
ordering and the fact we're setting this in the setState `updater` callback
a race condition is impossible here: if the condition
`st.mic_idx.rx_idx <= stash_idx` holds we really are putting the latest
response into the text field.